refactor(ZNTA-2297): ActivityTab and activity-util to typescript
