+++ server/zanata-frontend/src/app/editor/components/ActivityFeedItem/index.js
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /*
  * Copyright 2016, Red Hat, Inc. and individual contributors as indicated by the
  * @author tags. See the copyright.txt file in the distribution for a full
