+++ server/zanata-frontend/src/app/editor/utils/activity-util.ts
@@ -1,0 +1,42 @@
+import {tuple} from '../../utils/tuple'
+import {Status} from './phrase'
+export const ALL = 'all'
+export const COMMENTS = 'comments'
+export const UPDATES = 'updates'
+export const filterActivityTypes = tuple(ALL, COMMENTS, UPDATES)
+export type ActivityFilter = typeof filterActivityTypes[number]
+
+export interface User {
+  name: string
+}
+
+export const activityTypes = tuple('comment', 'revision')
+export type ActivityType = typeof activityTypes[number]
+
+export interface ActivityItem {
+  type: ActivityType,
+  content: string,
+  lastModifiedTime: Date,
+  status: Status,
+  user: User
+}
+
+export interface ActivityItemList extends Array<ActivityItem> {}
+
+export const activityItems = [
+  {
+    id: ALL,
+    icon: 'clock',
+    label: 'All'
+  },
+  {
+    id: COMMENTS,
+    icon: 'comment',
+    label: 'Comments'
+  },
+  {
+    id: UPDATES,
+    icon: 'refresh',
+    label: 'Updates'
+  }
+]
